# Conflicts:

#	leshan-server-demo/src/main/java/org/eclipse/leshan/server/demo/User.java

#	leshan-server-demo/src/main/java/org/eclipse/leshan/server/demo/servlet/BrokerServlet.java

+++ leshan-server-demo/src/main/java/org/eclipse/leshan/server/demo/User.java
@@ -1,20 +1,20 @@
 package org.eclipse.leshan.server.demo;
 
 public class User {
-	
+
 	public String UserID;
 	public int GroupNo;
 	public String RoomId;
 	public String Name;
 	public String Email;
 	private String Password;
-	private int locationX;
-	private int locationY;
+	private double locationX;
+	private double locationY;
 
 	private String LightUSER1;
 	private String LightUSER2;
 	private String Sensor;
-	
+
 	public User(String userID, int groupNo, String roomId, String name, String email, String password) {
         UserID = userID;
         GroupNo = groupNo;
@@ -26,8 +26,8 @@ public class User {
         LightUSER2 = "";
         Sensor = "";
     }
-	
-	public void setLocation(int X, int Y) {
+
+	public void setLocation(double X, double Y) {
 		locationX = X;
 		locationY = Y;
 	}
@@ -44,21 +44,21 @@ public class User {
 	public String getSensor() {
 		return Sensor;
 	}
-	
+
 	public String getLightUSER1() {
 		return LightUSER1;
 	}
 	public String getLightUSER2() {
 		return LightUSER2;
 	}
-	
-	
+
+
 	public double getDistanceFromLocation(int X, int Y) {
 		//sqrt((x2-x1)^2 + (y2-y1)^2)
-		double distance = Math.sqrt(Math.pow(X - locationX, 2) + Math.pow(Y - locationY, 2));		
+		double distance = Math.sqrt(Math.pow(X - locationX, 2) + Math.pow(Y - locationY, 2));
 		return distance;
 	}
-	
+
 	public Boolean checkPassword(String password){
 		return password.equals(Password);
 	}
